+++ engine/src/main/java/org/terasology/rendering/nui/layers/mainMenu/gameDetailsScreen/GameDetailsScreen.java
@@ -48,11 +48,11 @@ import org.terasology.rendering.nui.layers.mainMenu.SelectGameScreen;
 import org.terasology.rendering.nui.layers.mainMenu.savedGames.GameInfo;
 import org.terasology.rendering.nui.layouts.ScrollableArea;
 import org.terasology.rendering.nui.widgets.UIButton;
-import org.terasology.rendering.nui.widgets.UIImageSlideshow;
 import org.terasology.rendering.nui.widgets.UILabel;
 import org.terasology.rendering.nui.widgets.UIList;
 import org.terasology.rendering.nui.widgets.UITabBox;
 import org.terasology.rendering.nui.widgets.UIText;
+import org.terasology.rendering.nui.widgets.slideshow.UITimedImageSlideshow;
 import org.terasology.utilities.time.DateTimeHelper;
 import org.terasology.world.biomes.Biome;
 import org.terasology.world.biomes.BiomeManager;
@@ -482,8 +482,8 @@ public class GameDetailsScreen extends CoreScreenLayer {
         this.gameInfo = gameInfo;
     }
 
-    protected void setPreviewImage(TextureRegion texture) {
-        UIImageSlideshow slider = find("preview", UIImageSlideshow.class);
+    public void setPreviewImage(TextureRegion texture) {
+        UITimedImageSlideshow slider = find("preview", UITimedImageSlideshow.class);
         if (slider != null) {
             slider.addImage(texture);
         }
